requirement.txt
airtest==1.1.3
xlsxwriter==1.2.7
pyecharts==1.7.0
pyyaml==5.3
addict==2.1.3